app/requirements.txt
fastapi
uvicorn
python-telegram-bot
python-jose[cryptography]
passlib
sqlalchemy
bcrypt